Remove unnecessary casting.
When using input() to get, well input, from the user, this is automatically returned as a string type. No need to cast to string as was the case twice in the client.py file. If you expect the input to be integer, then you must cast to int but need to of course be careful that the user has indeed inputted an integer. I.e. error checking.handling is required!