Key exchange working!
Both client and server use a mixture of public and private variables to create the same private key. This is tested and working! Reduced the upper bound for the random to make the key exchange unnoticeable to the normal person With the larger upper bound, it took a second or two to compute the calculations involved.