Fixed server side and og client
The original client can now finish executing their actions without issue. The problem was that the second client with incorrect password was connecting to the server but not increasing the number of connections with that ID, however when it logged out due to the error, it decreased the number of connections with that ID, which resulted in an error in the original client when they tried to disconnect.

Also the connectionreseterror is caught and simply notifies the observer of the server that the client refused to connect and send a message.