Fixed the fernet error
Want to kms, This is a shite way to fix it but I cannot get to the bottom of this error otherwise. Now I know what it feels like when a detective cannot catch the fugitive on the run, this bug will haunt me for the rest of my days.